RemoteControlWebInterface: Auto download nodejs and use a dependecy (Windows only) #jria UE-161346 #rb Simon.Therriault
[CL 21594920 by Amir Keren in ue5-main branch]

+++ Engine/Plugins/VirtualProduction/RemoteControlWebInterface/WebApp/Clean.bat
@@ -16,7 +16,7 @@ goto :EOF
 :DEL_FOLDERS
 REM Using Robocopy instead of rmdir
 REM as node_modules may contains really long file path, and rmdir fails to delete them
-echo Deleting %~1
+REM echo Deleting %~1
 if not exist empty\ mkdir empty
 robocopy empty "%~1" /mir > nul
 rmdir /q /s "%~1"
@@ -29,6 +29,6 @@ REM Files are marked as readonly because of perforce
 REM Node v16 fails to install dependencies because it tries to overwrite the file
 REM if we force node to skip it, earlier versions of node will fail if we do that
 REM so we have original files prefixed with _, at build they are being copied and the readonly attribute is removed
-del /f /q "%~1\package.json"
-del /f /q "%~1\package-lock.json"
+if exist "%~1\package.json"         del /f /q "%~1\package.json"
+if exist "%~1\package-lock.json"    del /f /q "%~1\package-lock.json"
 goto :EOF
